Conflicts:
	src/Bomb.cpp
	src/Game.cpp

+++ include/Placement.hpp
@@ -9,11 +9,11 @@ class Placement
 private:
   int _numberPlayer;
   int _div;
-  glm::vec2 _mapSize;
+  glm::ivec2 _mapSize;
   Map *_map;
-  glm::vec2 const genNewPos();
-  bool checkSpace(glm::vec2 const &pos, glm::vec2 const &back, int rec);
-  glm::vec2 const foundCloserGoodPlace(glm::vec2 const &pos);
+  glm::ivec2 const genNewPos();
+  bool checkSpace(glm::ivec2 const &pos, glm::ivec2 const &back, int rec);
+  glm::ivec2 const foundCloserGoodPlace(glm::ivec2 const &pos);
 public:
   Placement(Map *_map);
   ~Placement();
